batch/requirements.txt
azure-batch==11.0.0
azure-storage-blob==12.8.1
azure-storage-file-datalake==12.8.1